genie simulator is enabled with prometheus metrics and renaming of Cerebral folders to genie
