--- agora/genie_mqtt_simulator/src/requirements.txt
@@ -1,1 +1,0 @@
-influxdb-client
