+++ agora/genie_simulator/src/requirements.txt
@@ -1,0 +1,3 @@
+influxdb-client
+prometheus_client
+psutil
